+++ documentation/stories/service/form.stories.jsx
@@ -47,10 +47,10 @@ ServiceCreation.args = {
   service: {},
   tenantName_id: [
     {
-      name: 'Demo1',
+      name: 'Tenant1',
       id: 'b2e35303-2747-4d1f-9767-519f9310b83e',
       props: {
-        name: 'Demo1',
+        name: 'Tenant1',
         icon: 'none',
         primaryColor: '#0018ef',
         secondaryColor: '#8086ba',
@@ -86,10 +86,10 @@ SubServiceCreation.args = {
   },
   tenantName_id: [
     {
-      name: 'Demo1',
+      name: 'Tenant1',
       id: 'b2e35303-2747-4d1f-9767-519f9310b83e',
       props: {
-        name: 'Demo1',
+        name: 'Tenant1',
         icon: 'none',
         primaryColor: '#0018ef',
         secondaryColor: '#8086ba',