+++ documentation/stories/shared/Card.stories.jsx
@@ -52,10 +52,10 @@ TenantExample.args = {
       title={'Title'}
       action={'modify'}
       tenant={{
-        name: 'Demo1',
+        name: 'Tenant1',
         id: 'b2e35303-2747-4d1f-9767-519f9310b83e',
         props: {
-          name: 'Demo1',
+          name: 'Tenant1',
           icon: 'none',
           primaryColor: '#0018ef',
           secondaryColor: '#8086ba',
@@ -76,10 +76,10 @@ TenantExample.args = {
     ></TenantForm>
   ),
   data: {
-    name: 'Demo1',
+    name: 'Tenant1',
     id: 'b2e35303-2747-4d1f-9767-519f9310b83e',
     props: {
-      name: 'Demo1',
+      name: 'Tenant1',
       icon: 'none',
       primaryColor: '#0018ef',
       secondaryColor: '#8086ba',
@@ -117,10 +117,10 @@ ServiceExample.args = {
       getServices={emptyFunction}
       tenantName_id={[
         {
-          name: 'Demo1',
+          name: 'Tenant1',
           id: 'b2e35303-2747-4d1f-9767-519f9310b83e',
           props: {
-            name: 'Demo1',
+            name: 'Tenant1',
             icon: 'none',
             primaryColor: '#0018ef',
             secondaryColor: '#8086ba',
@@ -151,5 +151,5 @@ ServiceExample.args = {
   colors: { secondaryColor: '#0018ef', primaryColor: '#0018ef' },
   getData: emptyFunction,
   seTenant: undefined,
-  tenantName_id: { name: 'Demo1', id: 'b2e35303-2747-4d1f-9767-519f9310b83e' }
+  tenantName_id: { name: 'Tenant1', id: 'b2e35303-2747-4d1f-9767-519f9310b83e' }
 };