+++ documentation/stories/shared/CardDialog.stories.jsx
@@ -57,10 +57,10 @@ MultifunctionButton.args = {
       getServices={emptyFunction}
       tenantName_id={[
         {
-          name: 'Demo1',
+          name: 'Tenant1',
           id: 'b2e35303-2747-4d1f-9767-519f9310b83e',
           props: {
-            name: 'Demo1',
+            name: 'Tenant1',
             icon: 'none',
             primaryColor: '#0018ef',
             secondaryColor: '#8086ba',