+++ RELEASE_NOTES.md
@@ -1,5 +1,21 @@
 # Auth Management UI Release Notes
 
+## 0.6
+
+### New features
+
+- Migration to realm v0.0.6
+
+### Bug fixes
+
+- Change the Group from name to path
+
+### Documentation
+
+### Continuous Integration
+
+### Technical debt
+
 ## 0.5
 
 ### New features